move everything in lib/submission_builder/shared up to the top level of submission_builder/
