+++ app/lib/submission_builder/ty2020/return1040.rb
@@ -1,6 +1,6 @@
 module SubmissionBuilder
   module Ty2020
-    class Return1040 < SubmissionBuilder::Shared::Return1040
+    class Return1040 < SubmissionBuilder::Return1040
       def attached_documents
         [{ xml_class: SubmissionBuilder::Ty2020::Documents::Irs1040 }]
       end
